feat:bug

+++ crmeb-common/src/main/java/com/zbkj/common/model/user/User.java
@@ -95,7 +95,7 @@ public class User implements Serializable {
     private BigDecimal brokeragePrice;
 
     @ApiModelProperty(value = "用户剩余积分")
-    private Integer integral;
+    private BigDecimal integral;
 
     @ApiModelProperty(value = "消费券")
     private BigDecimal consumeVoucher;
